pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cresus"
version = "0.1.0"
description = "Portfolio management application with FastAPI backend, MCP server, and React frontend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "cmd2>=2.4.0",
    "mcp>=1.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pandas>=2.2.0",
    "numpy>=1.24.0",
    "yfinance>=0.2.0",
    "gspread>=6.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "pyarrow>=12.0.0",
]

[project.scripts]
cresus = "cli.main:main"
cresus-api = "api.main:main"
cresus-mcp = "mcp.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages]
find = { where = ["src"] }
